cambios en la validacón de datos. Idioma de requests de error

+++ database/migrations/2018_07_06_183722_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             $table->string('nombre');
             $table->string('codigo');
             $table->integer('cantidad');
-            $table->integer('precio');
+            $table->double('precio', 8, 2);
             $table->timestamps();
         });
     }
